refactor: migrate socketErrorHandler to separate module
